refactor: rename original Context class to ConvesationContext to avoid naming conflicts
