+++ autogen/beta/annotations.py
@@ -8,7 +8,7 @@ from typing import Annotated, Any
 
 from fast_depends.library import CustomField
 
-from .context import Context as ContextType
+from .context import ConversationContext
 from .utils import CONTEXT_OPTION_NAME
 
 
@@ -93,6 +93,6 @@ class ContextField(CustomField):
 # `anything: Context`
 # are equal now
 Context = Annotated[
-    ContextType,
+    ConversationContext,
     ContextField(cast=False),
 ]
