+++ autogen/beta/config/anthropic/anthropic_client.py
@@ -17,7 +17,7 @@ from anthropic.types import (
 )
 
 from autogen.beta.config.client import LLMClient
-from autogen.beta.context import Context
+from autogen.beta.context import ConversationContext
 from autogen.beta.events import (
     BaseEvent,
     ModelMessage,
@@ -79,7 +79,7 @@ class AnthropicClient(LLMClient):
     async def __call__(
         self,
         messages: Sequence[BaseEvent],
-        context: Context,
+        context: "ConversationContext",
         *,
         tools: Iterable[ToolSchema],
         response_schema: ResponseProto | None,
@@ -169,7 +169,7 @@ class AnthropicClient(LLMClient):
     async def _process_response(
         self,
         response: Message,
-        context: Context,
+        context: "ConversationContext",
     ) -> ModelResponse:
         text_parts: list[str] = []
         calls: list[ToolCallEvent] = []
@@ -210,7 +210,7 @@ class AnthropicClient(LLMClient):
     async def _process_stream(
         self,
         stream: Any,
-        context: Context,
+        context: "ConversationContext",
     ) -> ModelResponse:
         full_content: str = ""
         calls: list[ToolCallEvent] = []
