+++ autogen/beta/config/client.py
@@ -5,7 +5,7 @@
 from collections.abc import Iterable, Sequence
 from typing import Protocol, runtime_checkable
 
-from autogen.beta.context import Context
+from autogen.beta.context import ConversationContext
 from autogen.beta.events import BaseEvent, ModelResponse
 from autogen.beta.response import ResponseProto
 from autogen.beta.tools.schemas import ToolSchema
@@ -16,7 +16,7 @@ class LLMClient(Protocol):
     async def __call__(
         self,
         messages: Sequence[BaseEvent],
-        context: Context,
+        context: ConversationContext,
         *,
         tools: Iterable[ToolSchema],
         response_schema: ResponseProto | None,
