+++ autogen/beta/observer.py
@@ -9,7 +9,7 @@ from typing import Any, Protocol, overload, runtime_checkable
 
 from autogen.beta.types import ClassInfo
 
-from .context import Context
+from .annotations import Context
 from .events.conditions import Condition, TypeCondition
 
 __all__ = ("Observer", "observer")
