+++ test/beta/config/anthropic/tools/test_mcp_server.py
@@ -5,8 +5,8 @@
 import pytest
 from dirty_equals import IsPartialDict
 
+from autogen.beta import Context
 from autogen.beta.config.anthropic.mappers import extract_mcp_servers, tool_to_api
-from autogen.beta.context import Context
 from autogen.beta.tools.builtin.mcp_server import MCPServerTool
