+++ test/beta/config/openai/tools/test_image_generation.py
@@ -4,8 +4,8 @@
 
 import pytest
 
+from autogen.beta import Context
 from autogen.beta.config.openai.mappers import tool_to_responses_api
-from autogen.beta.context import Context
 from autogen.beta.tools.builtin.image_generation import ImageGenerationTool
