+++ test/beta/config/openai/tools/test_mcp_server.py
@@ -4,8 +4,8 @@
 
 import pytest
 
+from autogen.beta import Context
 from autogen.beta.config.openai.mappers import tool_to_api, tool_to_responses_api
-from autogen.beta.context import Context
 from autogen.beta.exceptions import UnsupportedToolError
 from autogen.beta.tools.builtin.mcp_server import MCPServerTool
